+++ tests/MySQLSessionTest.php
@@ -4,7 +4,7 @@
 
    use Alo\Db\MySQL;
 
-   class SQLSessionTest extends \PHPUnit_Framework_TestCase {
+   class MySQLSessionTest extends \PHPUnit_Framework_TestCase {
 
       /**
        * @dataProvider definedProvider
@@ -59,7 +59,7 @@
             \Alo::$db = new MySQL('127.0.0.1', 3306, 'root', '', 'phpunit');
          }
 
-         if(!\Alo::$session || !(\Alo::$session instanceof SQLSession)) {
+         if(!\Alo::$session || !(\Alo::$session instanceof MySQLSessionTest)) {
             \Alo::$db->query('CREATE TABLE IF NOT EXISTS `alo_session` (
   `id`     CHAR(128)
            CHARACTER SET `ascii` NOT NULL,
@@ -71,7 +71,7 @@
   ENGINE =InnoDB
   DEFAULT CHARSET =`utf8mb4`;');
 
-            \Alo::$session = new SQLSession();
+            \Alo::$session = new MySQLSessionTest();
          }
 
          return \Alo::$session;